Fix some compile warnings which were caused by recent interface changes in file reader/writer.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5339 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CMS3DMeshFileLoader.cpp
@@ -159,7 +159,7 @@ bool CMS3DMeshFileLoader::load(io::IReadFile* file)
 	// read whole file
 
 	u8* buffer = new u8[fileSize];
-	s32 read = file->read(buffer, fileSize);
+	long read = (long)file->read(buffer, fileSize);
 	if (read != fileSize)
 	{
 		delete [] buffer;
